Add optional feature to get interlanguage links from wikibase
This requires setting $wgUnlinkedWikibaseSitelinkSuffix to the
wiki family (e.g. 'wiki' for wikipedia). If unset, no language
links are used. It also supports $wgUnlinkedWikibaseSitelinkSkippedLangs
if you want to skip some languages. For example, if your wiki is
english you might want to skip 'en'.

Badges are not supported

The language links use the unlinkedwikibase entity cache and
additionally are cached in parser cache with the page.

Bug: T355725
Change-Id: Ie0aeb3b7ffb77ce8c0f642f55c363f428f8fd398